+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
-## [2.0.0-beta6] - In progress
+## [2.0.0] - 2024-03-22
 
 ### Added
 
@@ -21,6 +21,7 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 - Force string cast in `HarGenerator` headers
 - Change method to get content and headers from CURL
 - Do not get content from HEAD requests
+- Bump version of `elgigi/har-parser` library to ^1.0
 
 ### Removed
 
@@ -33,6 +34,7 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 - Non-utf8 header value for HAR generation
 - Non-utf8 querystring value for HAR generation
 - Bug with clone method of PHP
+- Fix redirect for 307 and 308 http status
 
 ## [2.0.0-beta5] - 2022-09-07